microservices/indexer/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
openai==0.27.7
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pydantic==1.10.9
loguru==0.7.0
numpy==2.0.0
pandas==2.2.2
#pytest==7.4.0